+++ packages/failures/lib/failures.dart
@@ -1,6 +1,6 @@
 library;
 
-export 'src/annotations/failure_error_annotation.dart';
+export 'src/annotations/as_failure_annotation.dart';
 export 'src/failures.dart';
 export 'src/failures/dio_failure.dart';
 export 'src/failures/failure.dart';