+++ packages/failures/lib/src/annotations/as_failure_annotation.dart
@@ -5,8 +5,8 @@ import '/src/failures/failure.dart';
 /// An annotation used to specify that [Failure] subclass should be
 /// generated and annotated enum used as error type.
 @Target({TargetKind.enumType})
-class FailureError {
-  const FailureError({
+class AsFailure {
+  const AsFailure({
     this.name,
     this.type = .exception,
   });
@@ -15,4 +15,4 @@ class FailureError {
   final FailureType type;
 }
 
-const failureError = FailureError();
+const asFailure = AsFailure();