+++ packages/failures_builder/lib/src/failure_enum_generator.dart
@@ -8,7 +8,7 @@ class FailureEnumGenerator {
   });
 
   final EnumElement definingEnum;
-  final FailureError definingAnnotation;
+  final AsFailure definingAnnotation;
 
   String generate() {
     String namedConstructors = '';